+++ doc/ucsschool-changelog/changelog.rst
@@ -14,6 +14,20 @@ Changelog
   * New package for LUSD import (:uv:bug:`57547`).
 
 
+.. _changelog-ucsschool-2024-09-18:
+
+Released on 2024-09-17
+======================
+
+Source package *ucs-school-umc-computerroom* in version ``12.0.18``:
+
+* Keep the computer room responsive even if the Veyon Proxy is unstable.
+  This will ensure that exams can be started and finished, even if Veyon functionality does not work (:uv:bug:`57604`).
+
+Source package *ucs-school-veyon-client* in version ``2.0.6``:
+
+* Internal improvements (:uv:bug:`57604`).
+
 .. _changelog-ucsschool-2024-09-13:
 
 Released on 2024-09-13